Changing the AppStore initialization in favor of redux time travel
The Redux Dev-Tools (especially timetravel feature) does not work with a
plain inectable AppStore instance. To solve this, we provide a function
that creates the AppStore and use it in the angular2 bootstrap method to
delay the creation of the AppStore. This way the Redux Dev-Tools
chrome-extension have enough time to register/inject itself before the
store is created.
More information / explaination:
https://github.com/zalmoxisus/redux-devtools-extension/issues/44#issuecomment-177415058